Fix chat template tests, move `make_tokenizer` in tests out to common for reuse
